refac: split contracts into modules
