+++ contracts/lifecycle/src/errors.rs
@@ -1,0 +1,30 @@
+#![no_std]
+
+use soroban_sdk::contracterror;
+
+#[contracterror]
+#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
+#[repr(u32)]
+pub enum ContractError {
+    NoMaintenanceHistory = 1,
+    UnauthorizedEngineer = 2,
+    UnauthorizedAdmin = 3,
+    HistoryCapReached = 4,
+    AssetNotFound = 5,
+    NotInitialized = 6,
+    AlreadyInitialized = 7,
+    InvalidConfig = 8,
+    Paused = 9,
+    InvalidTaskType = 10,
+    PendingAdminAlreadyExists = 11,
+    ZeroAddress = 12,
+    SameRegistryAddress = 13,
+    IndexOutOfBounds = 14,
+    UnauthorizedOwner = 15,
+    TimelockNotExpired = 16,
+    ProposalNotFound = 17,
+    EngineerNotAuthorized = 16,
+    ScoreOverflow = 16,
+    /// Notes field exceeds the configured maximum length.
+    NotesTooLong = 18,
+}
